# lib/api-zod/src/generated/types/toggleUserActiveResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GESTAVICOLE SaaS API - Gestion avicole multi-tenant
 * OpenAPI spec version: 0.1.0
 */

export interface ToggleUserActiveResponse {
  id: string;
  isActive: boolean;
  message: string;
}